feat(@mikojs/logger): use new logger
